Event Collection Platform from social networks
The system supports two types of users,
the guests and the administrator .

Guest (guest)
The Guest to the home page, has access to a list of all events
registered in the system. This list is paged and sorted by event's tart
time (events that stats sooner appears first). For each event of the
list there are informations the user might need . Moreover the user is
able to search events by category,date (ability to set time) as well as
keywords that will sought in the description and the name of the event.
Last but not least, the environment supports mobile users (mobile
friendly).

Administrator (admin)
The administrator will have access to a system management interface
(control panel) in wich he can login in only with username and password.
From the administration interface, it can add and remove social network
resources (events).
The system retrieves and displays the details of the page (name,
picture, address) and allows the administrator database entry. In the
case of Facebook, we used Graph API, that allows a variety of ways in
order to (e.g., by JSON, HTTP Requests or via PHP SDKetc.) access an
event page.
The database will be updated with new events that create the registered
pages when the administrator clicks the corresponding operation.